refactor: UserNotFoundException 중복 익셉션 삭제
