+++ src/main/java/binbean/binbean_BE/constants/Constants.java
@@ -33,6 +33,8 @@ public class Constants {
 
     public static class ErrorMsg {
         public static final String USER_NOT_FOUND = "사용자를 찾을 수 없습니다.";
+        public static final String USER_NOT_FOUND_WITH_EMAIL = "사용자를 찾을 수 없습니다. (email: %s)";
+        public static final String USER_NOT_FOUND_WITH_ID = "사용자를 찾을 수 없습니다. (id: %s)";
         public static final String INVALID_CREDENTIALS = "이메일 또는 비밀번호가 올바르지 않습니다.";
         public static final String AUTHENTICATION_FAILED = "인증 실패";
         public static final String ACCESS_DENIED = "접근 권한이 없습니다.";
