+++ src/main/java/binbean/binbean_BE/exception/NotFoundException.java
@@ -1,8 +1,9 @@
 package binbean.binbean_BE.exception;
 
-public class NotFoundException extends RuntimeException {
+import org.springframework.http.HttpStatus;
 
+public class NotFoundException extends ClientErrorException {
     public NotFoundException(String message) {
-        super(message);
+        super(HttpStatus.NOT_FOUND, message);
     }
 }
